@@ -1,9 +1,13 @@
 <template>
-  <div class="ods-date">
+  <div class="ods-date" :class="{ 'ods-date--error': hasError }">
     <label class="ods-date__label">
       {{ label }}
-      <input type="text" class="ods-date__input" :value="displayDate" autocomplete="off" v-on:focus="toggleDatepicker(true)" />
+      <input type="text" class="ods-date__input" :value="displayDate" :placeholder="placeholder" autocomplete="off" v-on:focus="toggleDatepicker(true)" @keyup="handleKeyboardInput" />
     </label>
+    <div role="alert" class="ods-status-message ods-status-message--danger" v-if="hasError">
+      <h2 class="ods-status-message__heading"><span class="ods-status-message__icon ods-icon--error-hexa" aria-hidden="true"></span>{{ errorMessage }}</h2>
+    </div>
+
     <nrk-core-datepicker class="ods-date__datepicker" ref="datepicker" v-show="showDatepicker" :days="days" :months="months">
       <fieldset class="ods-date__datepicker__nav">
         <button class="ods-date__datepicker__button ods-date__datepicker__button--prev" :value="browseMonth(-1)" :disabled="browseMonthDisabled(-1)" :aria-label="btnPrevMonthLabel"></button>
@@ -22,6 +26,10 @@ window.customElements.define('nrk-core-datepicker', CoreDatepicker);
 export default {
   name: 'OdsDate',
   props: {
+    placeholder: {
+      type: String,
+      default: 'dd.mm.yyyy',
+    },
     label: {
       type: String,
       default: 'Pick a date',
@@ -58,13 +66,32 @@ export default {
       type: String,
       default: 'Next month',
     },
+    minDateErrorMessage: {
+      type: String,
+      default: 'Date is before the earliest allowed date.',
+    },
+    maxDateErrorMessage: {
+      type: String,
+      default: 'Date is after the latest allowed date',
+    },
+    invalidInputErrorMessage: {
+      type: String,
+      default: 'Invalid date or format. Try dd.mm.yyyy',
+    },
   },
 
   data: () => ({
     datepicker: null,
     showDatepicker: false,
     browseDate: new Date(Date.now()),
+    errorMessage: '',
   }),
+
+  computed: {
+    hasError() {
+      return this.errorMessage !== '';
+    },
+  },
 
   mounted() {
     this.datepicker = this.$refs.datepicker;
@@ -77,7 +104,7 @@ export default {
     window.addEventListener('keyup', this.onDatepickerOutside);
   },
 
-  destroyed() {
+  unmounted() {
     this.datepicker.removeEventListener('datepicker.change', this.onDatepickerChange);
     this.datepicker.removeEventListener('datepicker.click.day', this.onDatepickerClickDay);
     window.removeEventListener('click', this.onDatepickerOutside);
@@ -108,9 +135,41 @@ export default {
     onDatepickerClickDay() {
       this.$emit('set', this.datepicker.date);
       this.toggleDatepicker(false);
+      this.errorMessage = '';
     },
     onDatepickerChange(event) {
       this.browseDate = event.detail;
+    },
+    handleKeyboardInput(event) {
+      if (event.key !== 'Enter' || event.keyCode !== 13) {
+        return;
+      }
+      const [day, month, year] = event.target.value.split('.').map(Number);
+      const inputDate = new Date(year, month - 1, day);
+
+      if (!day || !month || !year || !this.isValidDate(inputDate)) {
+        this.errorMessage = this.invalidInputErrorMessage;
+        this.$emit('set', null);
+      } else if (inputDate < this.min) {
+        this.errorMessage = `${this.minDateErrorMessage} (${this.min.toLocaleDateString().replaceAll('/', '.')}).`;
+        this.$emit('set', null);
+      } else if (inputDate > this.max) {
+        this.errorMessage = `${this.maxDateErrorMessage} (${this.max.toLocaleDateString().replaceAll('/', '.')}).`;
+        this.$emit('set', null);
+      } else {
+        this.datepicker.date = inputDate;
+        this.$emit('set', this.datepicker.date);
+        this.toggleDatepicker(false);
+        this.errorMessage = '';
+      }
+    },
+
+    isValidDate(date) {
+      return !Number.isNaN(date.getTime());
+    },
+
+    isWithinRange(date) {
+      return date >= this.min && date <= this.max;
     },
   },
 };
